Making optimization block consistent
The import block for the fonts has now an unnecessary sign.
So it happened to the "attempt of optimization block".
Also, the code within that block is now consistent and optimized. I at least expect it to be 20% faster from now on (all changes were made on the run, so I forgot how it felt before doing so).

I'll do the same for the rest of blocks, now that the reversed engineering is going well so far.